Move the piece image array to its own file.
